doc(vanilla): write readme
