+++ packages/react/src/configurations.ts
@@ -9,9 +9,9 @@ export const configurationIdentityServer = {
     authority: 'https://demo.duendesoftware.com',
     // authority_time_cache_wellknowurl_in_second: 60* 60,
     refresh_time_before_tokens_expiration_in_second: 40,
-    service_worker_relative_url: '/OidcServiceWorker.js',
-    service_worker_only: false,
-    // storage: sessionStorage,
+    // service_worker_relative_url: '/OidcServiceWorker.js',
+    // service_worker_only: false,
+    storage: localStorage,
     // silent_login_timeout: 3333000
     // monitor_session: true,
     token_renew_mode: TokenRenewMode.access_token_invalid,
