+++ packages/react/src/oidc/User.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import { VanillaOidc } from './vanilla/vanillaOidc';
+import { OidcUserInfo, VanillaOidc } from './vanilla/vanillaOidc';
 
 export enum OidcUserStatus {
     Unauthenticated= 'Unauthenticated',
@@ -14,38 +14,6 @@ export type OidcUser<T extends OidcUserInfo = OidcUserInfo> = {
     status: OidcUserStatus;
 }
 
-export interface OidcUserInfo {
-    sub: string;
-    name?: string;
-    given_name?: string;
-    family_name?: string;
-    middle_name?: string;
-    nickname?: string;
-    preferred_username?: string;
-    profile?: string;
-    picture?: string;
-    website?: string;
-    email?: string;
-    email_verified?: boolean;
-    gender?: string;
-    birthdate?: string;
-    zoneinfo?: string;
-    locale?: string;
-    phone_number?: string;
-    phone_number_verified?: boolean;
-    address?: OidcAddressClaim;
-    updated_at?: number;
-}
-
-export interface OidcAddressClaim {
-    formatted?: string;
-    street_address?: string;
-    locality?: string;
-    region?: string;
-    postal_code?: string;
-    country?: string;
-}
-
 export const useOidcUser = <T extends OidcUserInfo = OidcUserInfo>(configurationName = 'default') => {
     const [oidcUser, setOidcUser] = useState<OidcUser<T>>({ user: null, status: OidcUserStatus.Unauthenticated });
 
@@ -57,6 +25,7 @@ export const useOidcUser = <T extends OidcUserInfo = OidcUserInfo>(configuration
             oidc.userInfoAsync()
                 .then((info) => {
                     if (isMounted) {
+                        // @ts-ignore
                         setOidcUser({ user: info, status: OidcUserStatus.Loaded });
                     }
                 })
